hack to fix Module did not self-register errors
it seems that removing .node references from the require cache leads to
issues with and error "Module did not self-register". Hence why if you
use bcrypt or anything you end up having to order your requires and
putting modules into your gulpfile that aren't needed in your gulp
script.

This is a temporary fix but it works
ref: sindresorhus/gulp-mocha #86