Stop swallowing errors in mocha output
If you copy paste the last example in the README to your source code,
then try to run mocha on a file that contains a ReferenceError, you
get no output. This makes it hard to track down the original cause of
the problem. Fix this by logging an error reported by Mocha to the
screen before exiting the process.